Record external node info in NormalizedMechanism
For the purposes of big Phi calculation, the causal structure of
mechanisms depends on which of their inputs and outputs lie outside of
the candidate set (which of them are "external nodes"). So, when
checking if a concept has already been computed (by looking up the hash
of a NormalizedMechanism) we need to take this into account; therefore
we record which of a mechanisms inputs/outputs are external when
normalizing a mechanism, and include this in its hash.