FIX: allow partitioned concepts to move to null
It turns out that in certain cases, in certain cases, the partitioned
system will have *greater* sum of small-phi, even though it has less
big-phi, which means that some partitioned constellation concepts will
need to be moved to the null concept (rather than only unpartitioned
concepts). This fix allows that to happen.