Add module for actual causation.
Primary object is a Context. It is built from a network, requires a
'before_state' and an 'after_state' as two sequential states of the
network. The context also requires two sets of indices, cause_indices
and effect_indices, which determine what elements are considered as
potential causes and effects when we cut the context.

An account is a set of all 'actions' for a context (actions are either
causes or effects, depending on direction) and it is calculated using
the 'directed_account' function. The irreducbility of a context is
tested by applying 'Actual_Cuts', and looking at the difference they
make to their combined forward and backward accounts.

A search over all sets of causes and effects is performed to find the
maximially irreducible context, the causal nexus.