Rework `Context` to use two `Subsystems`
Instead of representing a `Context` with one `Subsystem` and
having to switch back and forth between `before_state` and
`after_state` (which I eventually want to disallow entirely and
make `Subsystem` completely immutable), a `Context` now has two
`Subsystem` objects associated with it: one in `before_state`
used to compute effect coefficients, and one in `after_state` for
computing cause coefficients.

The API is, for now, identical.