Single nodes with selfloops have non-hardcoded phi.
Previously, when SINGLE_NODES_WITH_SELFLOOPS_HAVE_PHI was True, no
actual check for the presence of a selfloop was done, and a hardcoded
BigMip of Phi=0.5 was returned regardless of distance measure or
constellation.

This commit now takes the BigPhi of such systems to be distance between
its unpartitioned constellation and the null concept. You can think of this as
being a special case of the normal MIP calculation, where the
single-element's selfloop is allowed to be cut (normal system cuts do
not cut selfloops).

Because a single-node MIP now requires computation, its handling has
been moved from the "models" module to "compute."