Fix JSON deserialization
Objects computed in parallel processes have different ids, so
`id(obj)` is not necessarily equal for identical objects.
Similarly, different processes have different hash seeds, so
`Subsystem` hashes precomputed in different processes are not
equal for equivalent subsystems.

By using object hashes as JSON ids, and not precomputing the
Subsystem hash (which is rarely used), we guarantee that
identical objects are memoized when deserialized.