subsystem: NaNs in TPM for external nodes
Replace the columns in the `Subsystem` TPM that correspond to external
nodes with `np.NaN`. These values should never be used, so this helps
protect users from accidentally misinterpreting the Subsystem TPM
entries.

IMPORTANT: This requires copying the `Network` TPM (currently the
Subsystem TPM is just a view of the Network TPM). This will be slow and
expensive if the TPM is large.