Fix #38: configuration in subprocesses
Configuration options that had been directly assigned within the
interpreter were not respected in subprocesses spawned by `MapReduce`
instances because of a change in the system call used by
`multiprocessing`. In Python 3.8 and above, it's `spawn` instead of
`fork`, so the subprocess no longer has access to the parent's memory,
and a new configuration object is created. Now `MapReduce` will pass the
parent's config object to the subprocess, which loads it.

Note:
  - the issue still arises when parallelization is done by the user
    utside of PyPhi.
  - the issue only affects direct assignment to the config object; using
    the `pyphi_config.yml` file still works as intended in both cases.