Write config changes to disk and load on workers
When using Ray for parallelism, this fixes a bug where changes to the
config object were not respected in parallel processes on non-Unix
systems with Python>=3.8. In that version, the default method of
creating a process was changed from `fork` to `spawn`, so that child
processes no longer inherit the same config object.

Any changes to the config are written to a managed file on the disk.
When Ray starts a worker process and imports PyPhi, this file is loaded
(once). Note that any changes to the config made after PyPhi is imported
on a worker will not be respected by that worker.