Only compute potential purviews if not overriden
In `subsystem.find_mice`, computing potential purviews can be very
expensive in some situations, and if the user has provided a short
iterable of purviews, then computing the potential purviews is not worth
it. So, we simply use the user-provided purviews directly, allowing the
user to decide whether to filter out reducible purviews.